requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
lpips>=0.1.4
scipy>=1.11.0
tqdm>=4.66.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.35.0
accelerate>=0.24.0
clean-fid>=0.1.35
torch-fidelity>=0.3.0
opencv-python>=4.8.0
einops>=0.7.0
timm>=0.9.0
